requirements.txt
fastapi
uvicorn[standard]
lightgbm
pandas
numpy
scikit-learn
joblib
sqlalchemy
psycopg2-binary
PyJWT
python-multipart